+++ src/redturtle/prenotazioni/behaviors/booking_folder/notifications/email/adapters.py
@@ -28,6 +28,7 @@ class BookingTransitionEmailSender:
             If True, the message will be sent even if the email is not allowed
             (ie. for operator notifications)
         """
+
         message = self.message_adapter.message
 
         if force or getUtility(