+++ src/redturtle/prenotazioni/behaviors/booking_folder/notifications/sms/adapters.py
@@ -22,6 +22,7 @@ class BookingNotificationSender:
         self.request = request
 
     def send(self):
+
         if self.is_notification_allowed():
             # dont foget to write the history log about sending
             # self.write_message_to_booking_history(self.booking, self.message_adapter.message_history)