+++ src/redturtle/prenotazioni/restapi/serializers/adapters/prenotazione.py
@@ -5,6 +5,7 @@ from plone import api
 from plone.restapi.interfaces import IFieldSerializer
 from plone.restapi.interfaces import ISerializeToJson
 from plone.restapi.serializer.converters import json_compatible
+from Products.CMFCore.utils import getToolByName
 from zope.component import adapter
 from zope.component import getMultiAdapter
 from zope.i18n import translate
@@ -17,6 +18,19 @@ from redturtle.prenotazioni import logger
 from redturtle.prenotazioni.content.prenotazione import IPrenotazione
 from redturtle.prenotazioni.content.prenotazione_type import IPrenotazioneType
 from redturtle.prenotazioni.interfaces import ISerializeToPrenotazioneSearchableItem
+
+
+def get_booking_wf_state_title(context):
+    portal_workflow = getToolByName(context, "portal_workflow")
+    state = portal_workflow.getInfoFor(context, "review_state")
+
+    return translate(
+        msgid=state,
+        domain="plone",
+        target_language=getToolByName(
+            context, "portal_languages"
+        ).getPreferredLanguage(),
+    )
 
 
 @implementer(ISerializeToJson)
@@ -59,6 +73,7 @@ class PrenotazioneSerializer:
             booking_expiration_date = booking_expiration_date.replace(
                 booking_date.year, booking_date.month, booking_date.day
             )
+
         return {
             "@id": self.prenotazione.absolute_url(),
             "UID": self.prenotazione.UID(),
@@ -75,9 +90,7 @@ class PrenotazioneSerializer:
             "booking_date": json_compatible(booking_date),
             "booking_expiration_date": json_compatible(booking_expiration_date),
             "booking_status": status["review_state"],
-            "booking_status_label": translate(
-                status["review_state"], context=self.request
-            ),
+            "booking_status_label": get_booking_wf_state_title(self.prenotazione),
             "booking_type": self.prenotazione.booking_type,
             "booking_folder_uid": booking_folder.UID(),
             "vacation": self.prenotazione.isVacation(),
@@ -115,9 +128,7 @@ class PrenotazioneSearchableItemSerializer:
             # "booking_room": None,
             "booking_gate": self.prenotazione.gate,
             "booking_status": status["review_state"],
-            "booking_status_label": translate(
-                status["review_state"], context=self.request
-            ),
+            "booking_status_label": get_booking_wf_state_title(self.prenotazione),
             "booking_status_date": json_compatible(status["time"]),
             "booking_status_notes": status["comments"],
             "email": self.prenotazione.email,